QMainWindow tweaks for Windows.
Quit when the main window is closed on Windows.

Use the menuBar() in the central widget, which now floats as its own
visible top level window. I would have preferred to use the GLWidget as
the central widget but it breaks the osmlayer, for some reason.

Call ourselves "centralWidget" and save/restore our own size.